DDS-293 Removed irrelevant code from DDoc4J that isn't related to DDOC and validation

+++ ddoc4j/src/main/java/org/digidoc4j/ddoc/c14n/EntityParser_Entity.java
@@ -38,7 +38,7 @@ public class EntityParser_Entity {
     public int get_IntegerValue() {
 
         if (this.get_IsHexNumber()) {
-            return Convert.ToInt32(this.get_HexNumberBytes(), (int)0);
+            return Convert.ToInt32(this.get_HexNumberBytes(), 0);
         }
 
         return Convert.ToInt32(this.get_Text());
@@ -61,7 +61,7 @@ public class EntityParser_Entity {
 
     public byte[] get_HexNumberBytes()
     {
-        return Convert.FromHexString(this.get_Text().substring((int)1));
+        return Convert.FromHexString(this.get_Text().substring(1));
     }
 
     public static EntityParser_Entity Of(EntityParser_Fragment f) {
